assessment tool changes

+++ database/migrations/2023_02_21_125758_create_questions_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->enum('type', ['multiple_choice', 'open_ended']);
             $table->foreignId('assessment_tool_id')->constrained('assessment_tools')->cascadeOnDelete()->cascadeOnUpdate();
             $table->foreignId('assessment_group_id')->constrained('assessment_groups')->cascadeOnDelete()->cascadeOnUpdate();
-            $table->integer('point')->nullable();
+            $table->float('point')->nullable();
             $table->timestamps();
         });
     }
